moving lexical parser to other file
