+++ lexical_parser.h
@@ -1,0 +1,9 @@
+#ifndef LEXICAL_PARSER
+#define LEXICAL_PARSER
+
+#include <stdio.h>
+
+// Include the necessary definitions from lexical_parser.c
+#include "lexical_parser.c"
+
+#endif // LEXICAL_PARSER
